running predictions from linear regression
